src/mcpo_simple_server/requirements.txt
fastapi>=0.115.14
uvicorn>=0.35.0
pydantic>=2.11.7
sse-starlette>=1.6.5
loguru>=0.7.3
python-dotenv>=1.0.0
uuid>=1.30
python-jose[cryptography]
passlib[bcrypt]>=1.7.4
aiofiles>=24.1.0
jinja2>=3.1.5
uv>=0.7.0
uvx>=2.5.1
tinydb>=4.8.2
nosqlite>=0.0.3
dictdatabase>=2.5.1
mcp>=1.11.0
babel>=2.12.1